refactor(catalog): isolate publication and pointer reader
