+++ apps/backend/src/catalog/distribution/public/publication/dumpGeneration.ts
@@ -1,7 +1,7 @@
-import { unsafeRepeatableReadReadOnlyTransaction } from "../../../database/core";
-import { parsePublicOrigin } from "../../../shared/publicUrls";
-import type { BackendObservationScope } from "../../../observability/sentry";
-import { loadPublicCatalogSnapshotInExecutor } from "./snapshot";
+import { unsafeRepeatableReadReadOnlyTransaction } from "../../../../database/core";
+import { parsePublicOrigin } from "../../../../shared/publicUrls";
+import type { BackendObservationScope } from "../../../../observability/sentry";
+import { loadPublicCatalogSnapshotInExecutor } from "../snapshot";
 import {
   getCatalogDumpStorageConfig,
   writeCatalogDumpToS3,
