+++ apps/backend/src/catalog/distribution/public/publication/dumpStorage.ts
@@ -1,11 +1,10 @@
 import { createHash } from "node:crypto";
-import { GetObjectCommand, PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
+import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import {
   addBackendBreadcrumb,
   type BackendObservationScope,
-} from "../../../observability/sentry";
-import { HttpError } from "../../../shared/errors";
-import type { CatalogPublicSnapshot } from "../../types";
+} from "../../../../observability/sentry";
+import type { CatalogPublicSnapshot } from "../../../types";
 
 export type CatalogDumpStorageConfig = Readonly<{
   bucketName: string;
@@ -20,23 +19,10 @@ export type CatalogDumpWriteResult = Readonly<{
   byteLength: number;
 }>;
 
-/** Alias object naming the immutable artifact `GET /v1/catalog` redirects to. */
-export type CatalogDumpPointer = Readonly<{
-  objectKey: string;
-  url: string;
-  generatedAt: string;
-}>;
-
-export const catalogDumpPointerUnavailableCode = "CATALOG_DUMP_POINTER_UNAVAILABLE";
-
 const maxS3AttemptCount = 3;
 const catalogDumpObjectKeyPrefix = "catalog";
 const latestCatalogDumpObjectKey = `${catalogDumpObjectKeyPrefix}/latest.json`;
 const pointerCatalogDumpObjectKey = `${catalogDumpObjectKeyPrefix}/pointer.json`;
-const immutableCatalogDumpObjectKeyPattern = new RegExp(
-  `^${catalogDumpObjectKeyPrefix}/[0-9a-f]{64}\\.json$`,
-  "u",
-);
 const catalogDumpContentType = "application/json; charset=utf-8";
 export const immutableCatalogDumpCacheControl = "public, max-age=31536000, immutable";
 const revalidatedCatalogDumpCacheControl = "public, max-age=60";
@@ -168,103 +154,6 @@ async function putCatalogDumpObjectWithRetries(params: Readonly<{
     throw new Error(
       `Failed to write the public catalog dump to s3://${params.bucketName}/${params.objectKey}: ${formatCatalogDumpS3ErrorSummary(error)}`,
     );
-  }
-}
-
-function parseCatalogDumpPointerJson(
-  bodyText: string,
-  cdnBaseUrl: string,
-): CatalogDumpPointer {
-  const parsed: unknown = JSON.parse(bodyText);
-  if (typeof parsed !== "object" || parsed === null) {
-    throw new Error("Public catalog dump pointer must be a JSON object.");
-  }
-
-  const { objectKey, url, generatedAt } = parsed as Readonly<{
-    objectKey?: unknown;
-    url?: unknown;
-    generatedAt?: unknown;
-  }>;
-  // The route turns objectKey into a Location header, so it is held to the key
-  // shape the builder writes. Without this a control character in the key would
-  // reach header validation and fail the request as a 500 instead of this 503.
-  if (typeof objectKey !== "string" || !immutableCatalogDumpObjectKeyPattern.test(objectKey)) {
-    throw new Error(
-      `Public catalog dump pointer objectKey does not name an immutable ${catalogDumpObjectKeyPrefix} artifact.`,
-    );
-  }
-  if (typeof generatedAt !== "string" || generatedAt === "") {
-    throw new Error("Public catalog dump pointer is missing generatedAt.");
-  }
-  if (typeof url !== "string" || url === "") {
-    throw new Error("Public catalog dump pointer is missing url.");
-  }
-
-  // The route sends clients to this URL, so a pointer that names anything other
-  // than the configured distribution is treated as unreadable instead of being
-  // turned into an open redirect.
-  if (url !== `${cdnBaseUrl}/${objectKey}`) {
-    throw new Error(
-      `Public catalog dump pointer url does not name an object on ${cdnBaseUrl}.`,
-    );
-  }
-
-  return { objectKey, url, generatedAt };
-}
-
-function createCatalogDumpPointerUnavailableError(
-  config: CatalogDumpStorageConfig | null,
-  error: unknown,
-): HttpError {
-  const location = config === null
-    ? "public catalog dump storage"
-    : `s3://${config.bucketName}/${pointerCatalogDumpObjectKey}`;
-  return new HttpError(
-    503,
-    `Public catalog dump pointer is unavailable from ${location}: ${formatCatalogDumpS3ErrorSummary(error)}`,
-    catalogDumpPointerUnavailableCode,
-  );
-}
-
-/**
- * Reads the alias object naming the current immutable catalog artifact.
- *
- * `GET /v1/catalog` redirects to `url` instead of recomputing the snapshot, so
- * an unreadable pointer has to fail loudly. Recomputing per request is exactly
- * the load the artifact pipeline removed, and reintroducing it as a fallback
- * would bring back the API Gateway timeouts it was built to stop.
- */
-export async function loadCatalogDumpPointerFromS3(
-  observationScope: BackendObservationScope,
-): Promise<CatalogDumpPointer> {
-  let config: CatalogDumpStorageConfig | null = null;
-
-  try {
-    const resolvedConfig = getCatalogDumpStorageConfig();
-    config = resolvedConfig;
-    const response = await runCatalogDumpS3OperationWithRetries({
-      operation: "get_object",
-      observationScope,
-      bucketName: resolvedConfig.bucketName,
-      objectKey: pointerCatalogDumpObjectKey,
-      run: async () => getCatalogDumpS3Client().send(new GetObjectCommand({
-        Bucket: resolvedConfig.bucketName,
-        Key: pointerCatalogDumpObjectKey,
-      })),
-    });
-
-    if (response.Body === undefined) {
-      throw new Error(
-        `S3 returned an empty body for s3://${resolvedConfig.bucketName}/${pointerCatalogDumpObjectKey}`,
-      );
-    }
-
-    return parseCatalogDumpPointerJson(
-      await response.Body.transformToString(),
-      resolvedConfig.cdnBaseUrl,
-    );
-  } catch (error) {
-    throw createCatalogDumpPointerUnavailableError(config, error);
   }
 }
